+++ node/resolvers/catalog/index.ts
@@ -1,6 +1,17 @@
-import { NotFoundError, UserInputError } from '@vtex/api'
+import { NotFoundError, ResolverWarning, UserInputError } from '@vtex/api'
 import { all } from 'bluebird'
-import { compose, equals, find, head, last, path, prop, split, test } from 'ramda'
+import {
+  compose,
+  equals,
+  find,
+  head,
+  last,
+  path,
+  prop,
+  split,
+  test,
+  toLower,
+} from 'ramda'
 
 import { toSearchTerm } from '../../utils/ioMessage'
 import { resolvers as autocompleteResolvers } from './autocomplete'
@@ -17,18 +28,17 @@ import { resolvers as searchResolvers } from './search'
 import { resolvers as skuResolvers } from './sku'
 import { Slugify } from './slug'
 
-interface SearchArgs {
-  query: string
-  map: string
+interface SearchContext {
+  brand: string | null
+  category: string | null
+  contextKey: string
+}
+
+interface SearchContextParams {
+  brand: string
+  department: string
   category: string
-  specificationFilters: [string]
-  priceRange: string
-  collection: string
-  salesChannel: string
-  orderBy: string
-  from: number
-  to: number
-  hideUnavailableItems: boolean
+  subcategory: string
 }
 
 /**
@@ -54,7 +64,7 @@ const lastSegment = compose<string, string[], string>(
   split('/')
 )
 
-function findInTree(tree: any, values: any, index = 0): any {
+function findInTree(tree: Category[], values: string[], index = 0): any {
   for (const node of tree) {
     const slug = lastSegment(node.url)
     if (slug.toUpperCase() === values[index].toUpperCase()) {
@@ -69,7 +79,7 @@ function findInTree(tree: any, values: any, index = 0): any {
 /** Get Category metadata for the search/productSearch query
  *
  */
-const categoryMetaData = async (_: any, args: any, ctx: any) => {
+const categoryMetaData = async (_: any, args: ProductsArgs, ctx: any): Promise<Metadata> => {
   const { query } = args
   const category = findInTree(
     await queries.categories(_, { treeLevel: query.split('/').length }, ctx),
@@ -83,19 +93,21 @@ const categoryMetaData = async (_: any, args: any, ctx: any) => {
 /** Get brand metadata for the search/productSearch query
  *
  */
-const brandMetaData = async (_: any, args: any, ctx: any) => {
+const brandMetaData = async (_: any, args: ProductsArgs, ctx: any): Promise<Metadata> => {
   const brands = await queries.brands(_, { ...args }, ctx)
+  const brandName = toLower(last(args.query.split('/')) || '')
   const brand = find(
     compose(
-      equals(args.query.split('/').pop(-1)),
+      equals(brandName),
+      toLower,
       Slugify,
       prop('name') as any
     ),
     brands
-  )
+  ) || {}
   return {
-    metaTagDescription: path(['metaTagDescription'], brand as any),
-    titleTag: path(['title'], brand as any) || path(['name'], brand as any),
+    metaTagDescription: path(['metaTagDescription'], brand),
+    titleTag: path(['title'], brand) || path(['name'], brand),
   }
 }
 
@@ -106,14 +118,17 @@ const brandMetaData = async (_: any, args: any, ctx: any) => {
  * @param args
  * @param ctx
  */
-const searchMetaData = async (_: any, args: any, ctx: any) => {
+const searchMetaData = async (_: any, args: ProductsArgs, ctx: any) => {
   const { map } = args
-  const lastMap = map.split(',').pop(-1)
-  const meta =
-    lastMap === 'c'
-      ? await categoryMetaData(_, args, ctx)
-      : lastMap === 'b' && (await brandMetaData(_, args, ctx))
-  return meta
+  const lastMap = last(map.split(','))
+
+  if (lastMap === 'c') {
+    return categoryMetaData(_, args, ctx)
+  }
+  if (lastMap === 'b') {
+    return brandMetaData(_, args, ctx)
+  }
+  return { titleTag: null, metaTagDescription: null }
 }
 
 /** TODO: This method should be removed in the next major.
@@ -170,18 +185,22 @@ export const queries = {
     }
   },
 
-  facets: async (_: any, { facets, query, map }: any, ctx: Context) => {
+  facets: async (_: any, { facets, query, map, hideUnavailableItems }: FacetsArgs, ctx: Context) => {
     const {
       dataSources: { catalog },
       clients,
     } = ctx
     let result
     const translatedQuery = await translateToStoreDefaultLanguage(clients, query)
+    const segmentData = ctx.vtex.segment
+    const salesChannel = segmentData && segmentData.channel.toString() || ''
 
+    const unavailableString =
+       hideUnavailableItems ? `&fq=isAvailablePerSalesChannel_${salesChannel}:1` : ''
     if (facets) {
       result = await catalog.facets(facets)
     } else {
-      result = await catalog.facets(`${translatedQuery}?map=${map}`)
+      result = await catalog.facets(`${translatedQuery}?map=${map}${unavailableString}`)
     }
     result.queryArgs = {
       query: translatedQuery,
@@ -240,7 +259,7 @@ export const queries = {
     return catalog.products(args)
   },
 
-  productSearch: async (_: any, args: SearchArgs, ctx: Context) => {
+  productSearch: async (_: any, args: ProductsArgs, ctx: Context) => {
     const {
       dataSources: { catalog },
       clients,
@@ -252,7 +271,7 @@ export const queries = {
     }
     const products = await queries.products(_, translatedArgs, ctx)
     const recordsFiltered = await catalog.productsQuantity(translatedArgs)
-    const { titleTag, metaTagDescription }: any = await searchMetaData(
+    const { titleTag, metaTagDescription } = await searchMetaData(
       _,
       translatedArgs,
       ctx
@@ -265,11 +284,11 @@ export const queries = {
     }
   },
 
-  brand: async (_: any, args: any, { dataSources: { catalog } }: Context) => {
+  brand: async (_: any, { id }: {id?: number}, { dataSources: { catalog } }: Context) => {
     const brands = await catalog.brands()
     const brand = find(
       compose(
-        equals(args.id),
+        equals(id),
         prop('id') as any
       ),
       brands
@@ -285,13 +304,18 @@ export const queries = {
 
   category: async (
     _: any,
-    { id }: any,
+    { id }: { id?: number },
     { dataSources: { catalog } }: Context
-  ) => catalog.category(id),
+  ) => {
+    if (id == null) {
+      throw new ResolverWarning(`No category ID provided`)
+    }
+    return catalog.category(id)
+  },
 
   categories: async (
     _: any,
-    { treeLevel }: any,
+    { treeLevel }: { treeLevel: number },
     { dataSources: { catalog } }: Context
   ) => catalog.categories(treeLevel),
 
@@ -320,10 +344,10 @@ export const queries = {
 
   searchContextFromParams: async (
     _: any,
-    args: any,
+    args: SearchContextParams,
     { dataSources: { catalog } }: Context
   ) => {
-    const response = {
+    const response: SearchContext = {
       brand: null,
       category: null,
       contextKey: 'search',
@@ -331,41 +355,32 @@ export const queries = {
 
     if (args.brand) {
       const brands = await catalog.brands()
-      const found = brands.find(
-        (brand: any) => brand.isActive && Slugify(brand.name) === args.brand
-      )
-      response.brand = found && found.id
+      const found = brands.find(brand => brand.isActive && Slugify(brand.name) === args.brand)
+      response.brand = found ? found.id : null
     }
 
     if (args.department) {
       const departments = await catalog.categories(2)
-      let found: Category
+      let found
 
-      found = departments.find((department: any) =>
+      found = departments.find((department) =>
         department.url.endsWith(`/${args.department.toLowerCase()}`)
       )
       if (args.category && found) {
         found = found.children.find(category =>
           category.url.endsWith(`/${args.category.toLowerCase()}`)
-        ) as any
+        )
       }
 
       if (args.subcategory && found) {
         found = found.children.find(subcategory =>
           subcategory.url.endsWith(`/${args.subcategory.toLowerCase()}`)
-        ) as any
+        )
       }
 
-      response.category = found && (found.id as any)
+      response.category = found ? found.id : null
     }
 
     return response
   },
-}
-
-interface Category {
-  id: string
-  name: string
-  url: string
-  children: Category[]
 }